refactor: remove unused code and improve texture handling in raycasting
